+++ CHANGES.md
@@ -1,4 +1,5 @@
 #### [unreleased]
+* now strip CSS URL scheme to avoid mixed media errors from server
 
 #### 6.4.1 - 6.4.3 / 2020-01-28
 * test explicity for `$template instanceof \Tribe\Events\Views\V2\Template`, fixes bug when also using Events Tickets